requirements.txt
streamlit
groq
python-docx
pymupdf
